pyproject.toml
[tool.poetry]
name = "cyberbrain"
version = "0.0.0.2"
description = "A Cyberbrain that helps you remember everything."
authors = ["laike9m <laike9m@gmail.com>"]
license = "MIT"
readme="README.md"
homepage = "https://github.com/laike9m/Cyberbrain"
repository = "https://github.com/laike9m/Cyberbrain"
keywords = ["debug", "debugging", "debugger"]
include = [
    "LICENSE",
]
classifiers = [
    "Development Status :: 1 - Planning",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7"
]

[tool.poetry.dependencies]
python = ">= 3.6.7"
crayons = "^0.2.0"
astor = "^0.8.0"
deepdiff = "^4.0"
dataclasses = "^0.6.0"
graphviz = "^0.11.1"
absl-py = "^0.7.1"
black = {version = "*", allow-prereleases = true}
executing = "^0.3.2"
Pympler = "^0.8.0"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
tox = "^3.13"
tox-run-command = "^0.4.0"
tox-pyenv = "^1.1"
pyhamcrest = "^1.9"
astpretty = "^1.6"
pytype = "*"
pytest-xdist = "^1.30"
pylint = {git = "https://github.com/PyCQA/pylint.git"}

[tool.pylint.messages_control]
disable = """
bad-continuation,attribute-defined-outside-init,W0511,R0902,R0913,W0622,C0116,C0103,
R0903,W0631,W0212,W0603,R1710,R0201
"""

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
